requirements.txt
Django==1.11.10
Pillow==5.2.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
python-ldap==3.1.0
pytz==2018.5
